feat: Add single create bundle command
This handles both image and helm charts config files so can be used
to replace both of the individual subcommands, which are now deprecated.
